Small code & confusion cleanup
Addresses confusion when signatures are generated, the naming is not correct and it should instead be named as follows.

I have also automatically added in the offsetting with references as it is a huge inconvenience having to manually add on the offset when the code is there initially to do it, It doesnt make sense so this addresses that problem.